Refactored to merge helper classes and move it to an accesible location
